Починил сборку на Debian 11

+++ src/renderer_sdl.hpp
@@ -27,7 +27,7 @@ class Renderable {
   Renderable(const std::vector<float>* buffer, const char* vertex_shader,
              const char* pixel_shader)
       : shader{new Shader{vertex_shader, pixel_shader}},
-        num_indices_{buffer->size() / 3} {
+        num_indices_{int(buffer->size() / 3)} {
     glGenVertexArrays(1, &vao);
     unsigned int vbo;
     glGenBuffers(1, &vbo);
